+++ src/KlarnaExpressCheckout.php
@@ -1,6 +1,11 @@
 <?php
 namespace Krokedil\KlarnaExpressCheckout;
 
+use Krokedil\Klarna\Features;
+use Krokedil\Klarna\PluginFeatures;
+use Krokedil\KlarnaExpressCheckout\Api\Controllers\Notifications;
+use Krokedil\KlarnaExpressCheckout\Blocks\OneStepBlocksIntegration;
+
 defined( 'ABSPATH' ) || exit;
 
 /**
@@ -9,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  * @package Krokedil\KlarnaExpressCheckout
  */
 class KlarnaExpressCheckout {
-	const VERSION = '1.4.3';
+	const VERSION = '2.0.0';
 
 	/**
 	 * Reference to the Session class.
@@ -47,6 +52,13 @@ class KlarnaExpressCheckout {
 	private $settings;
 
 	/**
+	 * Reference to the WebhookSetup class.
+	 *
+	 * @var WebhookSetup
+	 */
+	private $webhook_setup;
+
+	/**
 	 * The ID of the payment button element.
 	 *
 	 * @var string
@@ -66,9 +78,15 @@ class KlarnaExpressCheckout {
 		$this->client_token_parser = new ClientTokenParser( $this->settings() );
 		$this->assets              = new Assets( $this->settings(), $locale );
 		$this->ajax                = new AJAX( $this->client_token_parser() );
+		$this->webhook_setup       = new WebhookSetup( $this );
 
 		add_action( 'init', array( $this, 'maybe_unhook_kp_actions' ), 15 );
 		add_action( 'woocommerce_single_product_summary', array( $this, 'add_kec_button' ), 31 );
+		add_action( 'woocommerce_blocks_loaded', array( $this, 'setup_blocks_integration' ) );
+        OneStepCheckout::register_hooks();
+
+		// Register the API controller for handling notifications in the Klarna API.
+		$this->register_api_controller();
 	}
 
 	/**
@@ -99,7 +117,11 @@ class KlarnaExpressCheckout {
 		// Get the instances of the Klarna Payments classes we need to unhook actions from.
 		$kp_session  = KP_WC()->session;
 		$kp_checkout = KP_WC()->checkout;
-		$kp_gateway  = WC()->payment_gateways()->get_available_payment_gateways()['klarna_payments'];
+		$kp_gateway  = WC()->payment_gateways()->get_available_payment_gateways()['klarna_payments'] ?? null;
+
+		if ( empty( $kp_session ) || empty( $kp_checkout ) || empty( $kp_gateway ) ) {
+			return;
+		}
 
 		// Unhook session actions.
 		remove_action( 'woocommerce_after_calculate_totals', array( $kp_session, 'get_session' ), 999999 );
@@ -159,6 +181,38 @@ class KlarnaExpressCheckout {
 	}
 
 	/**
+	 * Sets up the Blocks integration class.
+	 *
+	 * @return void
+	 */
+	public function setup_blocks_integration() {
+		// Only if One step is available.
+		if ( ! PluginFeatures::is_available( Features::KEC_ONE_STEP ) ) {
+			return;
+		}
+
+		if ( ! class_exists( 'Automattic\WooCommerce\Blocks\Package' ) || ! version_compare( \Automattic\WooCommerce\Blocks\Package::get_version(), '4.4.0', '>' ) ) {
+			return;
+		}
+		/**
+		 * Filter the compatible blocks for Klarna Express Checkout.
+		 */
+		$compatible_blocks = apply_filters(
+			'kec_compatible_blocks',
+			[ 'cart', 'mini-cart' ]
+		);
+
+		foreach ( $compatible_blocks as $block_name ) {
+			add_action(
+				"woocommerce_blocks_{$block_name}_block_registration",
+				function( $integration_registry ) {
+					$integration_registry->register( new OneStepBlocksIntegration() );
+				}
+			);
+		}
+	}
+
+	/**
 	 * Get the session class.
 	 *
 	 * @return Session
@@ -201,5 +255,23 @@ class KlarnaExpressCheckout {
 	 */
 	public function settings() {
 		return $this->settings;
+	}
+
+	/**
+	 * Get the webhook setup class.
+	 *
+	 * @return WebhookSetup
+	 */
+	public function webhook_setup() {
+		return $this->webhook_setup;
+	}
+
+	/**
+	 * Register the API controller for handling notifications in the Klarna API.
+	 *
+	 * @return void
+	 */
+	public function register_api_controller() {
+		Notifications::register_controller();
 	}
 }